chore(dynamic model): refactor class generation
The target of this refactor is the use of php attributes to mark properties as lazy.

The lazy information requires the use of a specific namespace.

Therefor the definition of used namespaces/classes needs to be managed.

Rendering the properties changed in order to place the correct information within the code.