fix(table loading): move add schema method into utility class
The method to add te dynamic SQL schema was removes from the dynamic model generator.

This previously leads to exceptions within the table analyzer as a result of other extensions (i.e. static info tables).

Since this method itself does not depend on methods or information of the dynamic model generator, this will not have any side effects.