chore(commands): strict types and readonly
The internal API for strict types and read only changed.
Previously these have been arguments to specific methods. These have now been removed.

Instead, corresponding method are used. This was a requirement for the command fourallportal:updateModels which did not support the strict option, since the corresponding method did not offer an argument for that.

This change should simplify the internal process and reduce the amount of arguments for methods.