fix(model generator): tca relation table configuration
The name of a relation table should not exceed 52 characters.
The use of the relation name that is delivered by the PIM is not unique threw out the system and the single column syntax has the same issue, only a combination of local and foreign column name will create a unique relation table name.

However, this combination will exceed 52 characters.

To solve this issue, table names that exceed 52 characters will be truncated to a maximum allowed characters and get a hash attached to it.

The use of configuration 'MM_opposite_field' has been corrected.