chore(dynamic model generator): improve relation calculation
Restore the use of static SQL definitions for relation tables.
Introduce new definition for relations with uid for table that allow multiple relations with the same data.

Note:
The calculation of the TCA needs more improvements.
The old behaviour has been restored, to missing information about the relation structure.

Log object and relation structure.
Cache configurations to reduce API calls.

The use of the dynamic model generator is not limited for one class.
To avoid that classes that are generated later in the process contain object stores of other classes, the cache variable needs to be reset before start generating a new class.

Remove columns related to workspaces.

Remove columns and relations tables that will be created by TYPO3 automatically.