chore(dynamic model): format skip reasons
The reason why a property was not process was displayed above the list of processed properties.

The listing of reasons did not contain the property or field name which makes is hard to identify what reason causes a field to be skipped.

The reasons are now collected within an array and print out as a table including the corresponding field.

This style makes it easier to identify the reason for each field.