chore(mapping): type binding for setter interface
The value setter interface now defines the type parameters in order. This change should take care, that the classes which are implement this interface knows exactly what kind of data they can expect.

The return type was set to void to indicate, that now return value was expected by the method.