chore: use column of type input for the TCA label
A label for the data set name should base on a value that can be displayed.

The previous implementation used the first column of the TCA which could not be guarantied to be a string.

This change determine the first column to be of type input and use it as the label for the data set.