fix(dynamic model): only use fields to load and relation fields
The process to calculate the fields used for a module previously did not perceive what fields are relevant and what relations belongs to the current model.

This process has been optimized.
Base of the calculation is now the information which fields should be used.

The relation information is checked and adds the correct fields with additional information about the relation style (self foreign) and if the field marks the main of child record.

Empty fields will be removed in the end of the merge process.

The logic how to build the TCA relation information has been stripped, since the information is now part of the field information.